# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification for Workout Tracker
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface Exercise {
  id: number;
  name: string;
  muscleGroup: string;
  /** bodybuilding | crossfit | cardio */
  category: string;
  /** @nullable */
  description?: string | null;
  createdAt: string;
}

export interface CreateExerciseBody {
  name: string;
  muscleGroup: string;
  category: string;
  /** @nullable */
  description?: string | null;
}

export interface Workout {
  id: number;
  name: string;
  /** bodybuilding | amrap | emom | rft | cardio */
  type: string;
  /** @nullable */
  description?: string | null;
  /**
   * duration in minutes (for AMRAP/EMOM)
   * @nullable
   */
  duration?: number | null;
  /**
   * number of rounds (for RFT)
   * @nullable
   */
  rounds?: number | null;
  /** JSON string of workout exercises/movements */
  exercises: string;
  createdAt: string;
}

export interface CreateWorkoutBody {
  name: string;
  type: string;
  /** @nullable */
  description?: string | null;
  /** @nullable */
  duration?: number | null;
  /** @nullable */
  rounds?: number | null;
  exercises: string;
}

export interface UpdateWorkoutBody {
  name?: string;
  type?: string;
  /** @nullable */
  description?: string | null;
  /** @nullable */
  duration?: number | null;
  /** @nullable */
  rounds?: number | null;
  exercises?: string;
}

export interface ScheduledWorkout {
  id: number;
  /** @nullable */
  workoutId?: number | null;
  workoutName: string;
  workoutType: string;
  scheduledDate: string;
  /** @nullable */
  notes?: string | null;
  completed: boolean;
  createdAt: string;
}

export interface CreateScheduledWorkoutBody {
  /** @nullable */
  workoutId?: number | null;
  workoutName: string;
  workoutType: string;
  scheduledDate: string;
  /** @nullable */
  notes?: string | null;
}

export interface UpdateScheduledWorkoutBody {
  /** @nullable */
  workoutId?: number | null;
  workoutName?: string;
  workoutType?: string;
  scheduledDate?: string;
  /** @nullable */
  notes?: string | null;
  completed?: boolean;
}

export interface WorkoutLog {
  id: number;
  /** @nullable */
  workoutId?: number | null;
  workoutName: string;
  workoutType: string;
  loggedAt: string;
  /** @nullable */
  durationMinutes?: number | null;
  /** @nullable */
  notes?: string | null;
  /** JSON string of results (sets/reps/weights or time/distance or rounds completed) */
  results: string;
  /**
   * 1-5 rating
   * @nullable
   */
  rating?: number | null;
  createdAt: string;
}

export interface CreateWorkoutLogBody {
  /** @nullable */
  workoutId?: number | null;
  workoutName: string;
  workoutType: string;
  loggedAt: string;
  /** @nullable */
  durationMinutes?: number | null;
  /** @nullable */
  notes?: string | null;
  results: string;
  /** @nullable */
  rating?: number | null;
}

export interface UpdateWorkoutLogBody {
  workoutName?: string;
  workoutType?: string;
  loggedAt?: string;
  /** @nullable */
  durationMinutes?: number | null;
  /** @nullable */
  notes?: string | null;
  results?: string;
  /** @nullable */
  rating?: number | null;
}

export type DashboardSummaryWorkoutsByType = {
  bodybuilding: number;
  amrap: number;
  emom: number;
  rft: number;
  cardio: number;
};

export interface DashboardSummary {
  totalWorkoutsLogged: number;
  workoutsThisWeek: number;
  workoutsThisMonth: number;
  scheduledThisWeek: number;
  totalMinutesThisMonth: number;
  workoutsByType: DashboardSummaryWorkoutsByType;
}

export interface WeeklyVolumeEntry {
  week: string;
  bodybuilding: number;
  amrap: number;
  emom: number;
  rft: number;
  cardio: number;
}

export type ListScheduledWorkoutsParams = {
  startDate?: string;
  endDate?: string;
};