# lib/api-zod/src/generated/types/createExerciseBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification for Workout Tracker
 * OpenAPI spec version: 0.1.0
 */

export interface CreateExerciseBody {
  name: string;
  muscleGroup: string;
  category: string;
  /** @nullable */
  description?: string | null;
}