# lib/api-zod/src/generated/types/dashboardSummary.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification for Workout Tracker
 * OpenAPI spec version: 0.1.0
 */
import type { DashboardSummaryWorkoutsByType } from "./dashboardSummaryWorkoutsByType";

export interface DashboardSummary {
  totalWorkoutsLogged: number;
  workoutsThisWeek: number;
  workoutsThisMonth: number;
  scheduledThisWeek: number;
  totalMinutesThisMonth: number;
  workoutsByType: DashboardSummaryWorkoutsByType;
}