# lib/api-zod/src/generated/types/exercise.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification for Workout Tracker
 * OpenAPI spec version: 0.1.0
 */

export interface Exercise {
  id: number;
  name: string;
  muscleGroup: string;
  /** bodybuilding | crossfit | cardio */
  category: string;
  /** @nullable */
  description?: string | null;
  createdAt: string;
}