# lib/api-zod/src/generated/types/scheduledWorkout.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification for Workout Tracker
 * OpenAPI spec version: 0.1.0
 */

export interface ScheduledWorkout {
  id: number;
  /** @nullable */
  workoutId?: number | null;
  workoutName: string;
  workoutType: string;
  scheduledDate: string;
  /** @nullable */
  notes?: string | null;
  completed: boolean;
  createdAt: string;
}