# lib/api-zod/src/generated/types/updateWorkoutBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification for Workout Tracker
 * OpenAPI spec version: 0.1.0
 */

export interface UpdateWorkoutBody {
  name?: string;
  type?: string;
  /** @nullable */
  description?: string | null;
  /** @nullable */
  duration?: number | null;
  /** @nullable */
  rounds?: number | null;
  exercises?: string;
}