# lib/api-zod/src/generated/types/updateWorkoutLogBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification for Workout Tracker
 * OpenAPI spec version: 0.1.0
 */

export interface UpdateWorkoutLogBody {
  workoutName?: string;
  workoutType?: string;
  loggedAt?: string;
  /** @nullable */
  durationMinutes?: number | null;
  /** @nullable */
  notes?: string | null;
  results?: string;
  /** @nullable */
  rating?: number | null;
}