# lib/api-zod/src/generated/types/workout.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification for Workout Tracker
 * OpenAPI spec version: 0.1.0
 */

export interface Workout {
  id: number;
  name: string;
  /** bodybuilding | amrap | emom | rft | cardio */
  type: string;
  /** @nullable */
  description?: string | null;
  /**
   * duration in minutes (for AMRAP/EMOM)
   * @nullable
   */
  duration?: number | null;
  /**
   * number of rounds (for RFT)
   * @nullable
   */
  rounds?: number | null;
  /** JSON string of workout exercises/movements */
  exercises: string;
  createdAt: string;
}